tools: validate HyperGeometric & DBSCAN against legacy on the local DB; fix DBSCAN epsilon type
Adds scripts/validation/validate_hypergeometric.py and validate_dbscan.py.

HyperGeometric: validated against a verbatim transcription of the legacy fisher() and an
independent scipy oracle over 45 gene-set pairs. Odds ratio matches legacy exactly (45/45);
upper tail matches legacy ut (45/45, the one tail before the accumulation bug); port
two-tailed matches scipy.stats.fisher_exact (45/45) and upper/lower match a scipy.stats
hypergeom pmf oracle (45/45). The legacy two-tailed is wrong in 35/45 pairs (the pval
accumulation bug the port fixes) -- so the port matches legacy where legacy is correct and
matches the oracle where it is not.

DBSCAN: both paths invoke the same dbscan C++ binary, so validation covers encode/gate/
decode. The bipartite input string is byte-identical to a verbatim legacy-encoding
transcription; across 6 (epsilon, min_points) settings on a real graph the ran gate, counts,
and decoded clusters all match (incl. a not-run case).

Bug fixed (found by validation): DBSCANInput.epsilon was typed float, so epsilon=1 produced
'1.0' which the binary's integer atol rejects ('Epsilon is invalid'). Epsilon is an integer
hop-radius for this binary -> typed int; tests updated. The fake-runner unit tests had used
epsilon=0.5 and never exercised the real binary.

Updates the migration doc (HyperGeometric/DBSCAN validation, epsilon fix, macOS build/fixture
notes). Tools unit suite: 74 passing.

+++ packages/tools/src/geneweaver/tools/dbscan/schema.py
@@ -14,7 +14,9 @@ class DBSCANInput(ToolInput):
     """
 
     gene_symbols: dict[str, list[str]] = Field(default_factory=dict)
-    epsilon: float
+    # Integer neighbourhood radius (BFS hop count). The dbscan binary parses epsilon with an
+    # integer ``atol`` (rejecting any non-digit, e.g. "1.0"), so this is an int, not a float.
+    epsilon: int
     min_points: int
     geneset_ids: list[str] = Field(default_factory=list)
